💫 UID Match

+++ src/DB/plan.js
@@ -1,14 +1,8 @@
-import { MongoClient, ServerApiVersion } from 'mongodb';
+import { MongoClient, ObjectId } from 'mongodb';
 
 const url = "mongodb+srv://Guynut:Guynut123@clusterplanmap.uwy5s2x.mongodb.net/?retryWrites=true&w=majority&appName=ClusterPlanMap";
 
-const client = new MongoClient(url, {
-  serverApi: {
-    version: ServerApiVersion.v1,
-    strict: true,
-    deprecationErrors: true,
-  }
-});
+const client = new MongoClient(url);
 
 let db;
 let collection;
@@ -25,42 +19,42 @@ export async function close() {
   console.log("Closed connection to MongoDB");
 }
 
-export async function createDocument(document) {
-  const result = await collection.insertOne(document);
+export async function createDocument(uid, document) {
+  const result = await collection.insertOne({ uid, ...document });
   console.log(`Document created with _id: ${result.insertedId}`);
   return result.insertedId;
 }
 
-export async function readDocument(documentId) {
-  const document = await collection.findOne({ _id: ObjectId(documentId) });
+export async function readDocument(uid, documentId) {
+  const document = await collection.findOne({ _id: ObjectId(documentId), uid });
   console.log(`Document read with _id: ${documentId}`);
   return document;
 }
 
-export async function readAllDocuments(sortField = "CreateAt", sortOrder = 1) {
+export async function readAllDocuments(uid, sortField = "CreateAt", sortOrder = 1) {
   const sortQuery = { [sortField]: sortOrder };
-  const documents = await collection.find({}).sort(sortQuery).toArray();
+  const documents = await collection.find({ uid }).sort(sortQuery).toArray();
   console.log("All documents read and sorted");
   return documents;
 }
 
-export async function updateDocument(documentId, update) {
+export async function updateDocument(uid, documentId, update) {
   const result = await collection.updateOne(
-    { _id: ObjectId(documentId) },
+    { _id: ObjectId(documentId), uid },
     { $set: update }
   );
   console.log(`Document updated with _id: ${documentId}`);
 }
 
-export async function deleteDocument(documentId) {
-  const result = await collection.deleteOne({ _id: ObjectId(documentId) });
+export async function deleteDocument(uid, documentId) {
+  const result = await collection.deleteOne({ _id: ObjectId(documentId), uid });
   console.log(`Document deleted with _id: ${documentId}`);
 }
 
-export async function findOneNearestToDate() {
+export async function findOneNearestToDate(uid) {
   const currentDate = new Date();
   const document = await collection
-    .find({})
+    .find({ uid })
     .sort({ $abs: { $subtract: ["$StartDate", currentDate] } })
     .limit(1)
     .next();
@@ -68,5 +62,3 @@ export async function findOneNearestToDate() {
   console.log("Nearest document to current date:", document);
   return document;
 }
-
-
